Add lenient word matching
More leniently match words similar in spelling. Words with differ by
minor spelling errors, American/British spelling differences,
tense/number, or common suffixes or prefixes to a target word would
still be considered (less significant) matches.

+++ templatefinder/src/main/java/database_templatefinder/templatefinder/InputProcessing.java
@@ -58,7 +58,7 @@ public class InputProcessing {
 		for(Entry<TemplateString, Double> entry : msgSet) {
 			count++;
 			if(count < msgSet.size() - 100) continue; // Only count the top 100
-			if(ps != null) ps.println(((int) (entry.getValue() * 1000000)) / 1000000.0 + " | " + entry.getKey().toImportantInfoString());
+			if(ps != null) ps.println(((long) (entry.getValue() * 1000000)) / 1000000.0 + " | " + entry.getKey().toImportantInfoString());
 		}
 		
 		if(ps != null) ps.println("Most likely template:   " + maxTemplate);
